testutil/integration: move app tests
