+++ .github/workflows/test.yml
@@ -27,7 +27,7 @@ jobs:
           token: ${{ secrets.CODECOV_TOKEN }}
           files: coverage.out
 
-  app_tests:
+  integration_tests:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -42,25 +42,7 @@ jobs:
           key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
           restore-keys: |
             ${{ runner.os }}-go-
-      - run: go test -v -timeout=5m -race github.com/obolnetwork/charon/app
-
-  app_docker_tests:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-go@v3
-        with:
-          go-version: '1.20.2'
-      - uses: actions/cache@v3
-        with:
-          path: |
-            ~/go/pkg/mod
-            ~/.cache/go-build
-          key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
-          restore-keys: |
-            ${{ runner.os }}-go-
-      - run: docker pull consensys/teku:latest
-      - run: go test -v -timeout=10m -race github.com/obolnetwork/charon/app -run=TestSimnetDuties -integration
+      - run: go test -v -timeout=10m -race github.com/obolnetwork/charon/testutil/integration -integration
 
   compose_tests:
     runs-on: ubuntu-latest
@@ -91,7 +73,7 @@ jobs:
 
   notify_failure:
     runs-on: ubuntu-latest
-    needs: [ unit_tests, app_tests, app_docker_tests, compose_tests ]
+    needs: [ unit_tests, integration_tests, compose_tests ]
     # Syntax ref: https://github.com/actions/runner/issues/1251
     if: always() && github.ref == 'refs/heads/main' && contains(join(needs.*.result, ','), 'failure')
     steps:
