+++ testutil/integration/doc.go
@@ -1,0 +1,7 @@
+// Copyright © 2022-2023 Obol Labs Inc. Licensed under the terms of a Business Source License 1.1
+
+// Package integration provides integration tests that includes multiple
+// instances communicating via loopback networking.
+//
+// Note all tests are skipped unless the -integration flag is provided.
+package integration
