+++ testutil/integration/simnet_test.go
@@ -1,12 +1,10 @@
 // Copyright © 2022-2023 Obol Labs Inc. Licensed under the terms of a Business Source License 1.1
 
-package app_test
+package integration_test
 
 import (
 	"context"
-	"flag"
 	"fmt"
-	"net"
 	"os"
 	"os/exec"
 	"path"
@@ -38,10 +36,11 @@ import (
 	"github.com/obolnetwork/charon/testutil/beaconmock"
 )
 
-//go:generate go test . -integration -v
-var integration = flag.Bool("integration", false, "Enable docker based integration test")
+//go:generate go test . -integration -v -run=TestSimnetDuties
 
 func TestSimnetDuties(t *testing.T) {
+	skipIfDisabled(t)
+
 	tests := []struct {
 		name                string
 		scheduledType       core.DutyType
@@ -123,11 +122,6 @@ func TestSimnetDuties(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			if test.teku && !*integration {
-				t.Skipf("Skipping Teku integration test (--integration=false): %v", t.Name())
-			} else if !test.teku && *integration {
-				t.Skipf("Skipping non-integration test (--integration=true): %v", t.Name())
-			}
 			t.Logf("Running test: %v", t.Name())
 
 			args := newSimnetArgs(t)
@@ -500,46 +494,4 @@ func startTeku(t *testing.T, args simnetArgs, node int) simnetArgs {
 	})
 
 	return args
-}
-
-// externalIP returns the hosts external IP.
-// Copied from https://stackoverflow.com/questions/23558425/how-do-i-get-the-local-ip-address-in-go.
-func externalIP(t *testing.T) string {
-	t.Helper()
-
-	ifaces, err := net.Interfaces()
-	require.NoError(t, err)
-
-	for _, iface := range ifaces {
-		if iface.Flags&net.FlagUp == 0 {
-			continue // interface down
-		}
-		if iface.Flags&net.FlagLoopback != 0 {
-			continue // loopback interface
-		}
-		addrs, err := iface.Addrs()
-		require.NoError(t, err)
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-			if ip == nil || ip.IsLoopback() {
-				continue
-			}
-			ip = ip.To4()
-			if ip == nil {
-				continue // not an ipv4 address
-			}
-
-			return ip.String()
-		}
-	}
-
-	t.Fatal("no network?")
-
-	return ""
 }
